+++ src/sig_gbbasic.h
@@ -1,0 +1,8 @@
+
+#ifndef _SIG_GBBASIC_H
+#define _SIG_GBBASIC_H
+
+void check_gbbasic(void);
+
+#endif // _SIG_GBBASIC_H
+