+++ src/sig_gbforth.h
@@ -1,0 +1,8 @@
+
+#ifndef _SIG_GBFORTH_H
+#define _SIG_GBFORTH_H
+
+void check_gbforth(void);
+
+#endif // _SIG_GBFORTH_H
+